+++ src/prommis/nanofiltration/diafiltration_flowsheet_two_salt.py
@@ -26,7 +26,7 @@ from idaes.models.unit_models import Feed, Product
 import matplotlib.pyplot as plt
 from pandas import DataFrame
 
-from prommis.nanofiltration.diafiltration_stream_properties import (
+from prommis.nanofiltration.multi_component_diafiltration_stream_properties import (
     DiafiltrationStreamParameter,
 )
 from prommis.nanofiltration.multi_component_diafiltration_solute_properties import SoluteParameter
